added duplicate and not found status code to accounts.get, post, delete

+++ api/sms/accounts.py
@@ -12,6 +12,8 @@ def get(username=None):
         users = [User.query.filter_by(username=username).first()]
     for user in UserSchema(many=True).dump(users):
         accounts.append(user)
+    if username and not user:
+        return accounts, 404
     return accounts, 200
 
 
@@ -20,12 +22,16 @@ def post(data):
     # TODO not recv this in plain-text
     hashed_password = bcrypt.generate_password_hash(data["password"]).decode("utf-8")
     data["password"] = hashed_password
+    if User.query.filter_by(username=data["username"]).all():
+        # username already taken
+        return None, 400
     new_user = UserSchema().load(data)
     db.session.add(new_user)
     try:
         db.session.commit()
     except:
-        return None, 500
+        # title already taken
+        return None, 400
     return None, 200
 
 
@@ -60,9 +66,8 @@ def manage(data):
 @accounts_decorator
 def delete(username):
     user = User.query.filter_by(username=username).first()
+    if not user:
+        return None, 404
     db.session.delete(user)
-    try:
-        db.session.commit()
-    except:
-        return None, 500
+    db.session.commit()
     return None, 200
